Split the call graph walkers into separate files.
